% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_files_accessible}
\alias{check_files_accessible}
\title{Check files}
\usage{
check_files_accessible(files)
}
\arguments{
\item{files}{character A vector of file names, no other limitation in length
than available memory to hold the data.}
}
\description{
Internal utility function used to check that file names passed
as argument are names as \code{character} and exist.
}
\keyword{internal}
