% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nearest_corners}
\alias{nearest_corners}
\alias{nearest_region}
\alias{nearest_rcoord}
\alias{nearest_ccoord}
\title{Nearest 2D grid region}
\usage{
nearest_corners(
  x,
  y,
  x.target,
  y.target,
  x.step = Inf,
  y.step = Inf,
  envelope = "nearest"
)

nearest_region(
  x,
  y,
  x.target,
  y.target,
  x.step = Inf,
  y.step = Inf,
  envelope = "nearest"
)

nearest_rcoord(
  x,
  y,
  x.target,
  y.target,
  x.step = Inf,
  y.step = Inf,
  envelope = "nearest"
)

nearest_ccoord(
  x,
  y,
  x.target,
  y.target,
  x.step = Inf,
  y.step = Inf,
  envelope = "nearest"
)
}
\arguments{
\item{x, y}{numeric vectors of equal length representing 2D coordinate pairs..}

\item{x.target, y.target}{numeric of length two giving the coordinates of the
corners of the target region.}

\item{x.step, y.step}{numeric of length one equal or larger than the largest
coordinate steps between grid points in \code{x} and \code{y}.}

\item{envelope}{character One of "nearest", "outer" or "inner".}
}
\value{
A vector of indices, assuming that \code{x} and \code{y} are coordinates
variables of data stored in long form.
}
\description{
Internal utility functions used to find indexes to rectangular
grid regions based on Euclidean distance.
}
\details{
The values of \code{x.step} and \code{y.step} are used to constrain the
computation of distances to the neighbourhood of the target. This is
to improve performance. The default \code{envelope = "outer"} returns the
coordinates of the smallest enclosing region.
}
\note{
For \code{"inner"} and \code{"outer"} envelopes the region is shrunk or expanded
by half a step. If the \code{step} is \code{Inf}, the largest steps found in \code{x} and
\code{y} replace \code{Inf}.
}
