% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nearest_point}
\alias{nearest_point}
\alias{nearest_pcoord}
\alias{dist2target}
\title{Nearest 2D grid point}
\usage{
nearest_point(x, y, x.target, y.target, x.step = Inf, y.step = Inf)

nearest_pcoord(x, y, x.target, y.target, x.step = Inf, y.step = Inf)

dist2target(x, y, x.target, y.target, x.step = Inf, y.step = Inf)
}
\arguments{
\item{x, y}{numeric vectors of equal length representing 2D coordinate pairs..}

\item{x.target, y.target}{numeric of length one giving the coordinates of the
nearest neighbor search.}

\item{x.step, y.step}{numeric of length one equal or larger than the largest
coordinate steps between grid points in \code{x} and \code{y}}
}
\value{
A numeric index, assuming that \code{x} and \code{y} are coordinates
variables of data stored in long form in the case of \code{nearest_point()}
or the values of \code{x} and \code{y} at this grid point in the case of
\code{nearest_pcoord()}.
}
\description{
Internal utility function used to find the nearest grid point
in a 2D coordinate system based on Euclidean distance.
}
\details{
The values of \code{x.step} and \code{y.step} are used to constrain the
computation of distances to the neighborhood of the target. This is
to improve performance.
}
