% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ac-saf-txt.R
\name{read_AC_SAF_txt}
\alias{read_AC_SAF_txt}
\alias{vars_AC_SAF_txt}
\alias{grid_AC_SAF_txt}
\title{Offline AC SAF Surface UV time series}
\usage{
read_AC_SAF_txt(
  files,
  vars.to.read = NULL,
  add.geo = length(files) > 1,
  keep.QC = TRUE,
  verbose = interactive()
)

vars_AC_SAF_txt(files, keep.QC = TRUE)

grid_AC_SAF_txt(files)
}
\arguments{
\item{files}{character A vector of file names, no other limitation in length
than available memory to hold the data.}

\item{vars.to.read}{character A vector of variable names. If `NULL` all the
variables present in the first file are read.}

\item{add.geo}{logical Add columns `Longitude` and `Latitude` to returned
data frame.}

\item{keep.QC}{logical Add to the returned data frame the quality control
variables, always present in the files.}

\item{verbose}{logical Flag indicating if progress, and time and size of
the returned object should be printed.}
}
\value{
A data frame with columns named `"Date"`, `"Longitude"`,
  `"Latitude"`, the data variables with their original names (with no units).
  The data variables have no metadata stored as R attributes. When reading
  multiple files, by default the format is similar to that from function
  `read_AC_SAF_hdf5()`. Column names are the same by column order can
  differ. File headers are saved as a list in R attribute `file.headers`.
}
\description{
All information is in the files, including dates, and no
  information is decoded from file names, that users will most likely want to
  rename. Each file corresponds to a single geographic location. If not all
  the files named in the argument to `files` are accessible, an error is
  triggered early. If the files differ in the coordinates, an error is
  triggered when reading the first mismatching file if coordinates are not
  being added to the data frame. Missing variables named in `vars.to.read`
  are currently ignored.

  Data from multiple files are concatenated. By default, the geographic
  coordinates are added in such a case.
}
\details{
Import time series "offline products" data files from EUMETSAT AC SAF
(Atmospheric Composition Monitoring) in text format. Currently only the
"Surface UV" data product files as downloaded from the FMI server are
supported.
}
\note{
When requesting the data from the EUMETSAT AC SAF FMI server at
  \url{https://acsaf.org/} it is possible to select the variables to be
  included in the file, the period and the geographic coordinates of a single
  location. The data are returned as a .zip compressed file containing one
  text file with one row for each day in the range of dates selected. These
  files are fairly small.

  This function's performance is not optimized for speed as these single
  location files are rather small. The example time series data included in
  the package are for one summer in Helsinki, Finland.
}
\examples{
# find location of one example file
one.file.name <-
   system.file("extdata", "AC_SAF_ts_Viikki.txt",
               package = "reumetsat", mustWork = TRUE)

# Available variables
vars_AC_SAF_txt(one.file.name)

# Grid point coordinates
grid_AC_SAF_txt(one.file.name)

# read all variables
summer_viikki.tb <-
  read_AC_SAF_txt(one.file.name)
dim(summer_viikki.tb)
colnames(summer_viikki.tb)
str(sapply(summer_viikki.tb, class))
summary(summer_viikki.tb)
attr(summer_viikki.tb, "file.headers")

# read all data variables
summer_viikki_QCf.tb <-
  read_AC_SAF_txt(one.file.name, keep.QC = FALSE)
dim(summer_viikki_QCf.tb)
summary(summer_viikki_QCf.tb)

# read all data variables including geographic coordinates
summer_viikki_geo.tb <-
  read_AC_SAF_txt(one.file.name, keep.QC = FALSE, add.geo = TRUE)
dim(summer_viikki_geo.tb)
summary(summer_viikki_geo.tb)

# read two variables
summer_viikki_2.tb <-
  read_AC_SAF_txt(one.file.name,
                    vars.to.read = c("DailyDoseUva", "DailyDoseUvb"))
dim(summer_viikki_2.tb)
summary(summer_viikki_2.tb)

}
\references{
Kujanpää, J. (2019) _PRODUCT USER MANUAL Offline UV Products v2
  (IDs: O3M-450 - O3M-464) and Data Record R1 (IDs: O3M-138 - O3M-152)_. Ref.
  SAF/AC/FMI/PUM/001. 18 pp. EUMETSAT AC SAFT.
}
