% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ac-saft-hdf5.R
\name{read_AC_SAFT_hdf5}
\alias{read_AC_SAFT_hdf5}
\title{Offline UV Products v2}
\usage{
read_AC_SAFT_hdf5(
  files,
  data.product = "Surface UV",
  group.name = "GRID_PRODUCT",
  vars_to_read = NULL,
  fill = NA_real_,
  verbose = getOption("verbose", default = interactive())
)
}
\arguments{
\item{files}{character A vector of file names, no other limitation in length
than available memory to hold the data.}

\item{data.product}{character Currently only "Surface UV" supported.}

\item{group.name}{character The name of the 'group' in the HDF5 files.}

\item{vars_to_read}{character A vector of variable names. If `NULL` all the
variables present in the first file are read.}

\item{fill}{numeric The R value used to replace the fill value used in the
file, which is retrieved from the file metadata.}

\item{verbose}{logical Flag indicating if progress, and time and size of
the returned object should be printed.}
}
\value{
A data frame with columns named `"Date"`, `"Longitude"`,
  `"Latitude"`, the data variables with their original names, and
  `"QualityFlags"`. The data variables have the metadata as attributes.
}
\description{
The dates are decoded from the file names, expecting these to be
  those set by the data provider. The grid is expected to be identical in all
  files that are imported in a call to `read_O3MOUV_hdf5()`. If not all the
  files named in the argument to `files` are accessible and error is
  triggered early. If the files differ in the grid, and error is triggered
  when reading the first mismatching file. Missing variables named in
  `vars_to_read` if detected when reading the first file are filled with the
  `fill` value, otherwise they trigger and error when an attempt is made to
  read them.
}
\details{
Import gridded daily "offline products" data files from EUMETSAT AC SAF
(Atmospheric Composition Monitoring) in HDF5 format.
}
\note{
When requesting the data from the EUMETSAT AC SAF the server it is
  possible to select the range of latitudes and longitudes and the variables
  to be included in the file. The data are returned as a .zip compressed file
  containing one HDF5 file for each day in the range selected. For world
  coverage each of these files can be as large as 10 MB in size. These are
  binary files so the size in RAM of a `data.frame` containing one-year of
  data can be a few 10's of GB. This function's performance is quite fast as
  long as there is enough RAM available and the files are read from a
  reasonably fast SSD.
}
\references{
\url{https://acsaf.org/}
}
