% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-omi-aura-uv-nc.R
\name{read_OMI_AURA_UV_nc}
\alias{read_OMI_AURA_UV_nc}
\alias{vars_OMI_AURA_UV_nc}
\alias{grid_OMI_AURA_UV_nc}
\alias{date_OMI_AURA_UV_nc}
\title{OMI/Aura Surface UV}
\usage{
read_OMI_AURA_UV_nc(
  files,
  vars.to.read = NULL,
  fill = NA_real_,
  verbose = interactive()
)

vars_OMI_AURA_UV_nc(files, set.oper = "intersect")

grid_OMI_AURA_UV_nc(files, expand = FALSE)

date_OMI_AURA_UV_nc(files, use.names = length(files > 1))
}
\arguments{
\item{files}{character A vector of file names, no other limitation in length
than available memory to hold the data.}

\item{vars.to.read}{character A vector of variable names. If \code{NULL} all the
variables present in the first file are read.}

\item{fill}{numeric The R value used to replace the fill value used in the
file, which is retrieved from the file metadata, and also used to fill
missing variables.}

\item{verbose}{logical Flag indicating if progress, and time and size of
the returned object should be printed.}

\item{set.oper}{character One of \code{"intersect"}, or \code{"union"}.}

\item{expand}{logical Flag indicating whether to return ranges or a
full grid.}

\item{use.names}{logical. Should names be added to the returned vector?}
}
\value{
Function \code{read_OMI_AURA_UV_nc()} returns a data frame with columns
named \code{"Date"}, \code{"Longitude"}, \code{"Latitude"}, and the data variables with their
original names. The data variables have their
metadata stored as R attributes. \code{vars_OMI_AURA_UV_nc()} returns a
\code{character} vector of variable names, \code{grid_OMI_AURA_UV_nc()} returns a
data frame with two numeric variables, \code{Longitude} and \code{Latitude}, with two
rows or an expanded grid depending on the argument passed to \code{expand},
while \code{date_OMI_AURA_UV_nc()} returns a named vector of class \code{Date}, with
file names as names.
}
\description{
Import \strong{gridded} "Surface UV" data released by
FMI/NASA from \strong{netcdf4} files downloaded from the NASA Earthwatch
server.
}
\details{
Function \code{read_OMI_AURA_UV_nc()} can be used to read the data
stored in a file, either in full or selected variables. Query functions
\code{vars_OMI_AURA_UV_nc()}, \code{grid_OMI_AURA_UV_nc()} and
\code{date_OMI_AURA_UV_nc()} extract the names of the variables, the range of
the grid and the dates of measurements more efficiently than by using
\code{read_OMI_AURA_UV_nc()}. The dates are decoded from the file names,
expecting these to be those set by the data provider or at least retaining
the date encoded as in the original name. The grid is expected
to be identical in all files that are imported in a call to
\code{read_OMI_AURA_UV_nc()}, and grid subsetting is currently not supported. If
not all the files named in the argument to \code{files} are accessible, an error
is triggered early. If the files differ in the grid, an error is triggered
when reading the first mismatching file. Missing variables named in
\code{vars.to.read} if detected when reading the first file, are filled with the
\code{fill} value, otherwise they trigger an error when an attempt is made to
read them.
}
\note{
The constraint on the consistency among all files to be read allows
very fast reading into a single data frame. If the files differ in the grid
or set of variables, this function can be used to read the files
individually into separate data frames. These data frames can later be
row-bound together.

The example data included in the package are only for Helsinki and
five summer days. They are used in examples and automated tests. Function
\code{read_OMI_AURA_UV_nc()} will be also tested by importing one-year's worth
of data with worldwide coverage on a PC with 64GB RAM.
}
\examples{
# find location of one example file
path.to.files <-
   system.file("extdata",
               package = "surfaceuv", mustWork = TRUE)

file.names <- list.files(path.to.files, pattern = "*.nc4$", full.names = TRUE)

# available variables
vars_OMI_AURA_UV_nc(file.names)

# available grid
grid_OMI_AURA_UV_nc(file.names)

# decode date from file name
date_OMI_AURA_UV_nc(file.names)
date_OMI_AURA_UV_nc(file.names, use.names = FALSE)

# read all variables
helsinki_3days.tb <- read_OMI_AURA_UV_nc(file.names)
dim(helsinki_3days.tb)
summary(helsinki_3days.tb)

}
