% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reumetsat-package.R
\docType{package}
\name{reumetsat-package}
\alias{reumetsat}
\alias{reumetsat-package}
\title{reumetsat: EUMETSAT Offline Data Products}
\description{
Functions for reading files downloaded from AC SAF EUMETSAT server containing data for offline products. Initially only Surface UV data supported.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.r4photobiology.info/reumetsat/}
  \item Report bugs at \url{https://github.com/aphalo/reumetsat/issues/}
}

}
\author{
\strong{Maintainer}: Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi} (\href{https://orcid.org/0000-0003-3385-972X}{ORCID})

}
\keyword{internal}
