% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reumetsat-package.R
\docType{package}
\name{surfaceuv-package}
\alias{surfaceuv}
\alias{surfaceuv-package}
\title{surfaceuv: Solar Ultraviolet at Ground Surface Data Import}
\description{
Functions for reading files containing data on "Surface UV" such as daily doses, irradiances and theUV Index (UVI). Curently the Surface UV product from the AC SAF project of EUMETSAT is supported in both geographic grid (HDF5) and time series (text) formats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.r4photobiology.info/surfaceuv/}
  \item Report bugs at \url{https://github.com/aphalo/surfaceuv/issues/}
}

}
\author{
\strong{Maintainer}: Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi} (\href{https://orcid.org/0000-0003-3385-972X}{ORCID})

}
\keyword{internal}
